CHALLENGE
I defined and learned how to utilize the HTML alt attribute.

Opened my Basic_HTML_1.html file and added an image (<img> element).

Utilized the alt attribute to specify an alternative text to be used when the image can’t be displayed.

Saved my code and successfully run it in the browser.

Returned to the Basic_HTML_1.html file and intentionally messed up the src attribute (for example: delete the “r” from src). Saved your code.

Ran your code in the browser. Due to the syntax error in the src attribute, the alternative text displays in the browser.

Lastly, I corrected the src attribute.